Config schema changed
Now each local repository must have .env file that contains variable
names to protect privacy and API keys that anyone might use. The config.py
is a generalized file that will load the variables that people want to use.
No variables will be stored in that file, but simply loaded in from the
.env file that contributors will have in their computer.